requirements.txt
annotated-types==0.7.0
asgiref==3.11.1
attrs==26.1.0
autobahn==25.12.2
Automat==25.4.16
cbor2==6.0.1
celery==5.4.0
certifi==2026.4.22
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.7
colorama==0.4.6
constantly==23.10.4
cryptography==47.0.0
daphne==4.2.1
dj-database-url==3.1.2
Django==5.2.13
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.71.2
gunicorn==25.3.0
httplib2==0.31.2
hyperlink==21.0.0
idna==3.13
Incremental==24.11.0
iniconfig==2.3.0
msgpack==1.1.2
packaging==26.2
pillow==12.2.0
pluggy==1.6.0
proto-plus==1.27.2
protobuf==5.29.6
psycopg==3.3.3
psycopg-binary==3.3.3
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
PyJWT==2.12.1
pyOpenSSL==26.1.0
pyparsing==3.3.2
PyPDF2==3.0.1
pytest==9.0.3
pytest-asyncio==1.3.0
pytest-django==4.12.0
python-dotenv==1.2.2
redis==7.4.0
reportlab==4.5.0
requests==2.33.1
service-identity==24.2.0
sqlparse==0.5.5
tqdm==4.67.3
Twisted==25.5.0
txaio==25.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
ujson==5.12.0
uritemplate==4.2.0
urllib3==2.6.3
whitenoise==6.12.0
zope.interface==8.4
